Enabling datatables to generate new style of control
